Make all id's type as Long

+++ app/src/main/java/kz/kazpost/loadingarea/api/_requests/AddSInvoicesToTInvoiceRequest.kt
@@ -7,7 +7,7 @@ data class AddSInvoicesToTInvoiceRequest(
     val sInvoiceIdsSeparatedByComma: String,
 
     @field:SerializedName("tlId")
-    val tInvoiceId: Int,
+    val tInvoiceId: Long,
 
     @field:SerializedName("averageVolume")
     val averageVolume: Long,
